moved some files and got rid of extra whitespace
