+++ src/main/java/gerstle/scrabble/Dictionary.java
@@ -15,18 +15,23 @@ public class Dictionary
      * @return list of words
      * @throws IOException for when the file is not found
      */
-    private final Map<String, String> wordsToDefinitions = new HashMap<>();
+    private final Map<String, String> wordsToDef = new HashMap<>();
 
     public Dictionary() throws IOException
     {
-        InputStream in = getClass().getClassLoader().getResourceAsStream("resources/dictionary.txt");
-        BufferedReader reader = new BufferedReader(new InputStreamReader(in));
-        String dictionaryLine;
-        while ((dictionaryLine = reader.readLine()) != null)
+        InputStream in = getClass().getClassLoader().getResourceAsStream("src/main/resources/dictionary.txt");
+        BufferedReader reader = null;
+        if (in != null)
         {
+            reader = new BufferedReader(new InputStreamReader(in));
+        }
+        String dictionaryLine;
+        while (reader.readLine() != null)
+        {
+            dictionaryLine = reader.readLine();
             int index = dictionaryLine.indexOf(" ");
             String[] pairs = dictionaryLine.split(" ", 2);
-            wordsToDefinitions.put(
+            wordsToDef.put(
                     index == -1 ? dictionaryLine : dictionaryLine.substring(0, index), //key
                     index > -1 ? dictionaryLine.substring(index + 1) : null);
         }
@@ -36,17 +41,17 @@ public class Dictionary
      */
     public boolean findWord(String word)
     {
-            return wordsToDefinitions.containsKey((word.toUpperCase()));
+            return wordsToDef.containsKey((word.toUpperCase()));
     }
 
     public String getDefinition(String word)
     {
-        String definition = wordsToDefinitions.get(word.toUpperCase());
+        String definition = wordsToDef.get(word.toUpperCase());
         return definition == null ? "" : definition;
     }
 
     public int size()
     {
-        return wordsToDefinitions.size();
+        return wordsToDef.size();
     }
 }
